pyproject.toml
[tool.poetry]
name = "electronics-network"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.8"
coverage = "^7.6.4"
django = "^5.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

